+++ lib/passes/CMakeLists.txt
@@ -1,19 +1,18 @@
 add_subdirectory(typegen)
 add_subdirectory(filter)
 add_subdirectory(analysis)
+add_subdirectory(configuration)
 
 set(PASS_SOURCES
   TypeARTPass.cpp
-  Commandline.cpp
   support/TypeUtil.cpp
   instrumentation/InstrumentationHelper.cpp
   instrumentation/TypeARTFunctions.cpp
   instrumentation/MemOpArgCollector.cpp
   instrumentation/MemOpInstrumentation.cpp
   instrumentation/Instrumentation.cpp
-  ../support/FileConfiguration.cpp
-  ../support/TypeARTOptions.cpp
   TypeARTConfiguration.cpp
+  Commandline.cpp
 )
 
 if(${LLVM_VERSION_MAJOR} VERSION_LESS_EQUAL "14")
@@ -27,6 +26,7 @@ typeart_make_llvm_module(
   LINK_LIBS typeart::MemInstFinder
             typeart::TypesStatic
             typeart::TypeGenDimetaStatic
+            typeart::PassConfiguration
             dimeta::Types
             ${TYPE_IR_GEN}
   INCLUDE_DIRS ${CMAKE_CURRENT_SOURCE_DIR}
