+++ lib/passes/configuration/Configuration.h
@@ -25,32 +25,6 @@
 
 namespace typeart::config {
 
-struct ConfigStdArgs final {
-#define TYPEART_CONFIG_OPTION(name, path, type, def_value, description, upper_path) static constexpr char name[] = path;
-#include "ConfigurationBaseOptions.h"
-#undef TYPEART_CONFIG_OPTION
-};
-
-struct ConfigStdArgValues final {
-#define TYPEART_CONFIG_OPTION(name, path, type, def_value, description, upper_path) \
-  static constexpr decltype(def_value) name{def_value};
-#include "ConfigurationBaseOptions.h"
-#undef TYPEART_CONFIG_OPTION
-};
-
-struct ConfigStdArgTypes final {
-#define TYPEART_CONFIG_OPTION(name, path, type, default_value, description, upper_path) using name##_ty = type;
-#include "ConfigurationBaseOptions.h"
-#undef TYPEART_CONFIG_OPTION
-};
-
-struct ConfigStdArgDescriptions final {
-#define TYPEART_CONFIG_OPTION(name, path, type, default_value, description, upper_path) \
-  static constexpr char name[] = description;
-#include "ConfigurationBaseOptions.h"
-#undef TYPEART_CONFIG_OPTION
-};
-
 namespace detail {
 template <typename T, typename Types>
 struct VariantContains;
